refactor: modularize conversion tracking types into separate files for improved organization and maintainability
